Place morse code table in flash address space
Modifies the morse code table's data structure so that it is an array of
fixed-length character arrays, rather than an array of pointers to string
literals.  The entire array is declared as const, which causes mspgcc to
put the whole thing in the .rodata section; now the table itself, and not
just its string literals, lives entirely in flash memory at runtime.